#34 - feat : [BE] 빌딩 정보 api, area_ready_status, delivery_area entitiy 생성

+++ backend/src/main/java/org/ssafy/load/dto/DeliveryArea.java
@@ -3,10 +3,10 @@ package org.ssafy.load.dto;
 import org.ssafy.load.domain.DeliveryAreaEntity;
 
 public record DeliveryArea(
-        Long id,
+        Integer id,
         String area_name
 ) {
-    public static DeliveryArea of(Long id, String area_name){
+    public static DeliveryArea of(Integer id, String area_name){
         return new DeliveryArea(id, area_name);
     }
